Corrección en tipo de dato

+++ src/routes/user_event.py
@@ -1,3 +1,4 @@
+from __future__ import annotations
 from fastapi import APIRouter, Depends
 from src.config.db import get_db
 from sqlalchemy.orm import Session
@@ -9,14 +10,15 @@ from src.schemas.query import QuerySchema
 from src.schemas.event import *
 from typing import List
 
+
 router = APIRouter(tags=["Events | User"])
 
 @router.get("/user/events", response_model=List[EventAllInfoSchemaOut], status_code=200)
 def get_events(title: str | None = None, 
                category: str | None = None, 
                direction: str  = "-", 
-               latitude: int  = 0, 
-               longitude: int = 0, 
+               latitude: float  = 0, 
+               longitude: float = 0, 
                offset: int = 0, 
                limit: int = 15, 
                user_db: User = Depends(verify),
